generate config from component data

+++ packages/anima-cli/src/commands/init-sb.ts
@@ -4,11 +4,11 @@ import * as Sentry from '@sentry/node';
 import fs from 'fs';
 import path from 'path';
 import { log } from '../helpers';
-import { generateStorybookConfig, getJSFiles, hasStorybook } from '../helpers/storybook';
+import { extractComponentInformation, generateStorybookConfig, getJSFiles, hasStorybook } from '../helpers/storybook';
 import { getToken } from '../helpers/token';
 import { authenticate } from '../api';
 import { execSync } from 'child_process';
-import { setDebug } from '../helpers/debug';
+import { isDebug, setDebug } from '../helpers/debug';
 
 export const command = 'init-sb';
 export const desc = 'Initialise storybook on your project';
@@ -34,12 +34,16 @@ export const handler = async (_argv: Arguments): Promise<void> => {
     setDebug(!!_argv.debug);
     const token = getToken(_argv);
     await authenticate(token);
-    loader.newStage('Install storybook');
-    const sampleStoriesPath = path.join('src', 'stories');
-    const hadStorybookFolderBefore = fs.existsSync(sampleStoriesPath);
-    execSync('npx storybook@latest init -y', {stdio: 'inherit'});
-    if(!hadStorybookFolderBefore && fs.existsSync(sampleStoriesPath)){
-      fs.rmSync(sampleStoriesPath, { recursive: true, force: true });
+    if(!fs.existsSync(".storybook")){
+      loader.newStage('Install storybook');
+      const sampleStoriesPath = path.join('src', 'stories');
+      const hadStorybookFolderBefore = fs.existsSync(sampleStoriesPath);
+      execSync('npx storybook@latest init -y', {stdio: 'inherit'});
+      if(!hadStorybookFolderBefore && fs.existsSync(sampleStoriesPath)){
+        fs.rmSync(sampleStoriesPath, { recursive: true, force: true });
+      }
+    } else {
+      console.log('Skipping storybook install');
     }
     loader.newStage('Fetching components');
     const componentsDir = _argv.components as string || 'src';
@@ -49,10 +53,13 @@ export const handler = async (_argv: Arguments): Promise<void> => {
     for(const componentFile of componentsWithoutStorybook){
       console.log(`Creating ${componentFile}...`);
       const componentContent = fs.readFileSync(componentFile, 'utf8');
-      const response = await generateStorybookConfig(componentContent, token).catch((e) => {throw e});
+      const response = await extractComponentInformation(componentContent, token).catch((e) => {throw e});
       if(response){
-        fs.writeFileSync(componentFile, response.code);
-        fs.writeFileSync(`${componentFile.split(".")[0]}.stories.js`, response.story);
+        if(isDebug()){
+          fs.writeFileSync(`${componentFile.split(".")[0]}.stories.log`, response.reply);
+        }
+        const storybookConfig = generateStorybookConfig(componentFile, response);
+        fs.writeFileSync(`${componentFile.split(".")[0]}.stories.js`, storybookConfig);
       } else {
         console.log(`Skipped ${componentFile}. Couldn't generate story config`);
       }
